requirements.txt
python-dotenv
google-generativeai
flask
requests
assemblyai